New version
Changed quite a few things...it utomatically loads calibration file based on date of run, it knows the structure of the MIRAGE folder as long as you point it to MIRAGE and it can analyse a single shot, a range of shots or an entire run. The file rossFilterExample shows how to call the function. As it is now it cannot be used with the datapipeline, but it kind of has it's own very primitive pipeline so it should work to just call the function...I think